+++ packages/spec/src/cloud/index.ts
@@ -21,5 +21,4 @@ export * from './environment.zod';
 export * from './environment-package.zod';
 export * from './package.zod';
 export * from './package-version.zod';
-export * from './project.zod';
 export * from './project-artifact.zod';